Add files via upload
This program utilizes the Iris Dataset, which contains four features of three Iris flower species. Using the iris dataset, this program classifies the data as the correct species of flower through the decision tree model. The output is a PDF file containing a visualization of the decision tree, as well as the specific iris species sorted by color.